fixup issues with 32-bit jvm
If you run tests under a 32-bit jvm, you will get a test failure in IndexStoreTests,
the logic there is wrong in the case of 32-bit (its NIOFSDirectory on linux).

Also if mlockall fails, you'll see huge bogus values (because of use of `long` instead of `NativeLong`)

finally add seccomp support for 32 bit too, and clean up all its `long` usage as well.

+++ core/src/main/java/org/elasticsearch/bootstrap/JNANatives.java
@@ -71,8 +71,8 @@ class JNANatives {
                 JNACLibrary.Rlimit rlimit = new JNACLibrary.Rlimit();
                 if (JNACLibrary.getrlimit(JNACLibrary.RLIMIT_MEMLOCK, rlimit) == 0) {
                     rlimitSuccess = true;
-                    softLimit = rlimit.rlim_cur;
-                    hardLimit = rlimit.rlim_max;
+                    softLimit = rlimit.rlim_cur.longValue();
+                    hardLimit = rlimit.rlim_max.longValue();
                 } else {
                     logger.warn("Unable to retrieve resource limits: " + JNACLibrary.strerror(Native.getLastError()));
                 }
